pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tervyx-evidence"
version = "0.1.0"
description = "TERVYX Evidence Extraction Pipeline (C Repo) - LLM-powered literature search, matching, and extraction for deterministic policy-as-code evaluation"
authors = [
    {name = "TERVYX Team", email = "contact@tervyx.org"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.19.0",
    "pydantic>=2.4.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "biopython>=1.81",  # For PubMed/Entrez
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-requests",
    "types-PyYAML",
]

llm = [
    "openai>=1.3.0",
    "anthropic>=0.7.0",
]

[project.scripts]
tervyx-evidence = "src.common.cli:main"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
